% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPeekeR.R
\name{peakPeekeR}
\alias{peakPeekeR}
\title{peakPeekeR}
\usage{
peakPeekeR(trt_bam, ctrl_bam = NULL)
}
\value{
A Shiny app for peak calling parameter optimization.
}
\description{
Does the peak peeking.
}
\author{
Jared Andrews, Lawryn Kasper, Jacob Steele
}
